# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "factfit",
    "description": "Tailor CVs without inventing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "summary": "Create Job",
        "description": "Paste a job description: it is parsed into requirements (cached per text).",
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJob"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/match": {
      "post": {
        "summary": "Match",
        "description": "Check the profile against a parsed job: status and evidence per requirement.",
        "operationId": "match_jobs__job_id__match_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MatchRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/jobs/{job_id}/tailor": {
      "post": {
        "summary": "Start Tailor",
        "description": "Start tailoring a CV for a parsed job; it pauses when drafts are ready for review.",
        "operationId": "start_tailor_jobs__job_id__tailor_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TailorRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "description": "Where a run is, with its drafts once they are ready for review.",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/review": {
      "post": {
        "summary": "Review",
        "description": "Accept, edit or reject each draft. Edits are checked; a failing edit comes back.",
        "operationId": "review_runs__run_id__review_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusOut"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/events": {
      "get": {
        "summary": "Run Events",
        "description": "Server-sent events: one per finished step, then the run's status.",
        "operationId": "run_events_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateJob": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text",
            "description": "The job description, pasted as text"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CreateJob"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobDescription": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "seniority": {
            "type": "string",
            "enum": [
              "intern",
              "fresher",
              "junior",
              "mid",
              "senior",
              "unknown"
            ],
            "title": "Seniority"
          },
          "language": {
            "type": "string",
            "enum": [
              "vi",
              "en"
            ],
            "title": "Language"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "must_have": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "title": "Must Have"
          },
          "nice_to_have": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "title": "Nice To Have"
          },
          "responsibilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Responsibilities"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "company",
          "seniority",
          "language",
          "location",
          "must_have",
          "nice_to_have",
          "responsibilities",
          "keywords"
        ],
        "title": "JobDescription"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "jd": {
            "$ref": "#/components/schemas/JobDescription"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "company",
          "cached",
          "jd"
        ],
        "title": "JobOut"
      },
      "MatchOut": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "intern",
                  "fresher",
                  "junior",
                  "mid",
                  "senior",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "profile_version": {
            "type": "string",
            "title": "Profile Version"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementOut"
            },
            "type": "array",
            "title": "Requirements"
          },
          "excluded": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "score",
          "level",
          "profile_version",
          "cached",
          "requirements",
          "excluded"
        ],
        "title": "MatchOut"
      },
      "MatchRequest": {
        "properties": {
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "intern",
                  "fresher",
                  "junior",
                  "mid",
                  "senior",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "Applicant level for postings that hire several; default: lowest accepted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MatchRequest"
      },
...
      "RequirementOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "bucket": {
            "type": "string",
            "enum": [
              "must",
              "nice"
            ],
            "title": "Bucket"
          },
          "category": {
            "type": "string",
            "enum": [
              "skill",
              "experience",
              "education",
              "language",
              "soft_skill",
              "other"
            ],
            "title": "Category"
          },
          "any_of": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Any Of"
          },
          "status": {
            "type": "string",
            "enum": [
              "met",
              "partial",
              "missing",
              "unverifiable"
            ],
            "title": "Status"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "bucket",
          "category",
          "any_of",
          "status",
          "note",
          "evidence"
        ],
        "title": "RequirementOut"
      },
...
      "ReviewRequest": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/ReviewDecisionIn"
            },
            "type": "array",
            "title": "Decisions",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReviewRequest"
      },
      "RunOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "waiting_review",
              "done",
              "error",
              "not_found"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunOut"
      },
      "RunStatusOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "waiting_review",
              "done",
              "error",
              "not_found"
            ],
            "title": "Status"
          },
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/DraftOut"
            },
            "type": "array",
            "title": "Drafts",
            "default": []
          },
          "review_errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Review Errors",
            "default": {}
          },
          "cv": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cv"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunStatusOut"
      },
      "TailorRequest": {
        "properties": {
          "level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "intern",
                  "fresher",
                  "junior",
                  "mid",
                  "senior",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "use_judge": {
            "type": "boolean",
            "title": "Use Judge",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TailorRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}